web-app/requirements.txt
Flask~=1.1
mysql-connector-python~=8.0
pandas~=1.3
matplotlib~=3.5
geopy~=2.2
pymemcache~=3.5
numpy~=1.21
